+++ apps/gitness/src/layouts/PullRequestLayout.tsx
@@ -1,54 +1,22 @@
 import React, { useEffect, useState } from 'react'
-import { NavLink, Outlet, useMatch, useParams } from 'react-router-dom'
+import { NavLink, Outlet, useParams } from 'react-router-dom'
 import { Badge, Icon, Spacer, Tabs, TabsList, TabsTrigger } from '@harnessio/canary'
 import { Floating1ColumnLayout, PullRequestHeader } from '@harnessio/playground'
 import { TypesPullReq, useGetPullReqQuery } from '@harnessio/code-service-client'
 import { useGetRepoRef } from '../framework/hooks/useGetRepoPath'
-import { PathParams, PullRequestRoutePathParams, routes } from '../RouteDefinitions'
-
-enum PullRequestTab {
-  CONVERSATION = 'conversation',
-  COMMITS = 'commits',
-  CHANGES = 'changes',
-  CHECKS = 'checks'
-}
+import { PathParams } from '../RouteDefinitions'
+import useGetPullRequestTab, { PullRequestTab } from '../hooks/useGetPullRequestTab'
 
 const PullRequestLayout: React.FC = () => {
   const [pullRequest, setPullRequest] = useState<TypesPullReq>()
-  const { spaceId, repoId } = useParams<PathParams>()
+  const { spaceId, repoId, pullRequestId } = useParams<PathParams>()
   const repoRef = useGetRepoRef()
-  const { pullRequestId } = useParams<PathParams>()
   const prId = (pullRequestId && Number(pullRequestId)) || -1
   const { data: pullRequestData, isFetching } = useGetPullReqQuery({
     repo_ref: repoRef,
     pullreq_number: prId
   })
-  const [pullRequestTab, setPullRequestTab] = useState<PullRequestTab | null>(null)
-  const urlMatchArgs: PullRequestRoutePathParams = {
-    spaceId: spaceId || '',
-    repoId: repoId || '',
-    pullRequestId: String(prId)
-  }
-  const routeTabMapping = [
-    {
-      match: useMatch(routes.toPullRequest(urlMatchArgs)),
-      tab: PullRequestTab.CONVERSATION
-    },
-    {
-      match: useMatch(routes.toPullRequestConversation(urlMatchArgs)),
-      tab: PullRequestTab.CONVERSATION
-    },
-    { match: useMatch(routes.toPullRequestCommits(urlMatchArgs)), tab: PullRequestTab.COMMITS },
-    { match: useMatch(routes.toPullRequestChanges(urlMatchArgs)), tab: PullRequestTab.CHANGES },
-    { match: useMatch(routes.toPullRequestChecks(urlMatchArgs)), tab: PullRequestTab.CHECKS }
-  ]
-
-  useEffect(() => {
-    const matchedRoute = routeTabMapping.find(route => route.match)
-    if (matchedRoute) {
-      setPullRequestTab(matchedRoute.tab)
-    }
-  }, [routeTabMapping, spaceId, repoId, prId])
+  const pullRequestTab = useGetPullRequestTab({ spaceId, repoId, pullRequestId })
 
   useEffect(() => {
     if (!isFetching && pullRequestData) {
@@ -60,7 +28,7 @@ const PullRequestLayout: React.FC = () => {
       <Floating1ColumnLayout>
         <Spacer size={8} />
         {pullRequest && <PullRequestHeader data={pullRequest} />}
-        <Tabs variant="tabnav" defaultValue={PullRequestTab.CONVERSATION} value={pullRequestTab}>
+        <Tabs variant="tabnav" value={pullRequestTab?.valueOf()}>
           <TabsList>
             <NavLink to={`conversation`}>
               <TabsTrigger value={PullRequestTab.CONVERSATION}>
